backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
openai-whisper
pydantic==2.9.2
python-dotenv==1.0.1
aiofiles==24.1.0
soundfile
